refactor(web): extract file upload logic into useFileUpload hook
Extract file upload handling from ChatPanel.tsx (1,816 lines) into
a dedicated useFileUpload hook (366 lines), reducing the component
to 1,592 lines.

The hook encapsulates:
- File validation (size, type, dimensions)
- Drag and drop handling
- Clipboard paste support
- File preview URL management
- Text file content reading
- Error state management

This improves code organization and makes the file upload logic
reusable in other components.
